Run PolicyKit + Metagov integration tests with GitHub workflow
